Add styling to answers delete button, add confirm message on delete answer.

+++ controllers/question.js
@@ -10,12 +10,15 @@ MessageBoard.QuestionController = Ember.ObjectController.extend({
     },
 
     deleteAnswer: function(answer) {
-     var question = this.get('model');
-     question.get('answers').removeObject(answer);
-     question.save();
-     answer.deleteRecord();
-     answer.save();
-     this.transitionToRoute('question', this.get('id'));
+
+     if (confirm('Are you sure you want to delete this answer?')){
+       var question = this.get('model');
+       question.get('answers').removeObject(answer);
+       question.save();
+       answer.deleteRecord();
+       answer.save();
+       this.transitionToRoute('question', this.get('id'));
+     }
    },
